リクエストの国ごとにメッセージ変える機能を１つのクラスにまとめた
